+++ simpleml/pipelines/validation_split_mixins.py
@@ -1,5 +1,5 @@
 '''
-Module for different split methods for cross validation
+Module for different pipeline split methods for cross validation
 
     1) No Split -- Just use all the data
     2) Explicit Split -- dataset class defines the split
@@ -15,64 +15,11 @@ import pandas as pd
 from abc import ABCMeta, abstractmethod
 from sklearn.model_selection import train_test_split
 from future.utils import with_metaclass
-from collections import defaultdict
-from typing import Any, Dict, List, Optional, Union
+from typing import Dict, List, Optional, Union
 
 from simpleml.constants import TRAIN_SPLIT, VALIDATION_SPLIT, TEST_SPLIT
-
-
-class Split(dict):
-    '''
-    Container class for splits
-    '''
-
-    def __getattr__(self, attr):
-        '''
-        Default attribute processor
-        (Used in combination with __getitem__ to enable ** syntax)
-        '''
-        return self.get(attr, None)
-
-    @staticmethod
-    def is_null_type(obj: Any) -> bool:
-        '''
-        Helper to check for nulls - useful to not pass "empty" attributes
-        so defaults of None will get returned downstream instead
-        ex: **split -> all non null named params
-        '''
-        # NoneType
-        if obj is None:
-            return True
-
-        # Pandas objects
-        if isinstance(obj, (pd.DataFrame, pd.Series)) and obj.empty:
-            return True
-
-        # Empty built-ins - uses __nonzero__
-        if isinstance(obj, (list, tuple, dict)) and not obj:
-            return True
-
-        # Else
-        return False
-
-    def squeeze(self):
-        '''
-        Helper method to clear up any null-type keys
-        '''
-        poppable_keys = [k for k, v in self.items() if self.is_null_type(v)]
-        [self.pop(k) for k in poppable_keys]
-
-        # Return self for easy chaining
-        return self
-
-
-class SplitContainer(defaultdict):
-    '''
-    Explicit instantiation of a defaultdict returning split objects
-    '''
-
-    def __init__(self, default_factory=Split, **kwargs):
-        super(SplitContainer, self).__init__(default_factory, **kwargs)
+from simpleml.datasets.dataset_splits import Split, SplitContainer
+from .projected_splits import IdentityProjectedDatasetSplit, IndexBasedProjectedDatasetSplit
 
 
 class SplitMixin(with_metaclass(ABCMeta, object)):
@@ -98,7 +45,7 @@ class NoSplitMixin(SplitMixin):
         '''
         Non-split mixin class. Returns full dataset for any split name
         '''
-        default_split = Split(X=self.dataset.X, y=self.dataset.y).squeeze()
+        default_split = IdentityProjectedDatasetSplit(dataset=self.dataset, split=None)
         self._dataset_splits = self.containerize_split({
             'default_factory': lambda: default_split
         })
@@ -108,18 +55,23 @@ class ExplicitSplitMixin(SplitMixin):
     def split_dataset(self) -> None:
         '''
         Method to split the dataframe into different sets. Assumes dataset
-        explicitly delineates between train, validation, and test
+        explicitly delineates between different splits
+
+        Passes forward dataset split names so uniquely named splits will propagate
+        and can be referenced the same way
         '''
         self._dataset_splits = self.containerize_split({
-            TRAIN_SPLIT: Split(X=self.dataset.get('X', TRAIN_SPLIT), y=self.dataset.get('y', TRAIN_SPLIT)).squeeze(),
-            VALIDATION_SPLIT: Split(X=self.dataset.get('X', VALIDATION_SPLIT), y=self.dataset.get('y', VALIDATION_SPLIT)).squeeze(),
-            TEST_SPLIT: Split(X=self.dataset.get('X', TEST_SPLIT), y=self.dataset.get('y', TEST_SPLIT)).squeeze()
+            split_name: IdentityProjectedDatasetSplit(dataset=self.dataset, split=split_name)
+            for split_name in self.dataset.get_split_names()
         })
 
 
 class RandomSplitMixin(SplitMixin):
     '''
     Class to randomly split dataset into different sets
+
+    **Redefines splits so custom named splits in dataset cannot be referenced
+    by the same names. Only TRAIN/TEST/VALIDATION**
     '''
 
     def __init__(self,
@@ -150,6 +102,19 @@ class RandomSplitMixin(SplitMixin):
             'shuffle': shuffle
         })
 
+    @staticmethod
+    def get_index(data) -> List[int]:
+        '''
+        Helper to extract the index from a dataset. Generates a range index
+        if none exists
+        '''
+        if isinstance(data, (pd.DataFrame, pd.Series)):
+            return data.index
+
+        else:
+            # no named index, use a linear range
+            return range(len(data))
+
     def split_dataset(self) -> None:
         '''
         Overwrite method to split by percentage
@@ -161,25 +126,29 @@ class RandomSplitMixin(SplitMixin):
         shuffle = self.config.get('shuffle')
 
         # Sklearn's train test split can only accomodate one split per iteration
+        # find the indices that match to each split
+        # use the X split section
+        index = self.get_index(self.dataset.X)
+
         if test_size == 0:  # No split necessary
-            X_remaining, y_remaining = self.dataset.X, self.dataset.y
-            X_test, y_test = [], []
+            test_indices = []
+            remaining_indices = index
         else:
-            X_remaining, X_test, y_remaining, y_test = train_test_split(
-                self.dataset.X, self.dataset.y, test_size=test_size, random_state=random_state, shuffle=shuffle)
+            remaining_indices, test_indices = train_test_split(
+                index, test_size=test_size, random_state=random_state, shuffle=shuffle)
 
         calibrated_validation_size = float(validation_size) / (validation_size + train_size)
         if calibrated_validation_size == 0:  # No split necessary
-            X_train, y_train = X_remaining, y_remaining
-            X_val, y_val = [], []
+            train_indices = remaining_indices
+            validation_indices = []
         else:
-            X_train, X_val, y_train, y_val = train_test_split(
-                X_remaining, y_remaining, test_size=calibrated_validation_size, random_state=random_state, shuffle=shuffle)
+            train_indices, validation_indices = train_test_split(
+                remaining_indices, test_size=calibrated_validation_size, random_state=random_state, shuffle=shuffle)
 
         self._dataset_splits = self.containerize_split({
-            TRAIN_SPLIT: Split(X=X_train, y=y_train).squeeze(),
-            VALIDATION_SPLIT: Split(X=X_val, y=y_val).squeeze(),
-            TEST_SPLIT: Split(X=X_test, y=y_test).squeeze()
+            TRAIN_SPLIT: IndexBasedProjectedDatasetSplit(dataset=self.dataset, split=None, indices=train_indices),
+            VALIDATION_SPLIT: IndexBasedProjectedDatasetSplit(dataset=self.dataset, split=None, indices=validation_indices),
+            TEST_SPLIT: IndexBasedProjectedDatasetSplit(dataset=self.dataset, split=None, indices=test_indices)
         })