+++ CHANGELOG.md
@@ -2,9 +2,15 @@
 
 ### TBD
 - Added support to hasher for initialized objects
+- Adds support for arbitrary dataset splits and sections
+- Dataset hooks to validate dataframe setting
+- Pipelines no longer cache dataset splits and proxy directly to dataset on every call
+- Introduces pipeline splits as reproducible projections over dataset splits
+- Database utility to recalculate hashes for existing persistables
 
-*Breaking Changes*
+*BREAKING CHANGES*
 - Hash for an uninitialized class changed from repr(cls) to "cls.__module_.cls.__name_"
+- Database migrations no longer recalculate hashes. That has to be done manually via a utility
 
 ### 0.10.0 (2021-07-09)
 - Dataset external file setter with validation hooks